+++ Views/Shared/_Layout.cshtml
@@ -33,7 +33,7 @@
                     <li><a asp-area="" asp-controller="Home" asp-action="Index">Home</a></li>
                     @if (User.Identity.IsAuthenticated)
                     {
-                        <li><a asp-area="" asp-controller="ShortUrls" asp-action="Index">Shorten new URL</a></li>
+                        <li><a asp-area="" asp-controller="ShortUrls" asp-action="Create">Shorten new URL</a></li>
                     }
                     else
                     {
