+++ Views/ShortUrls/Show.cshtml
@@ -34,8 +34,8 @@
             <span class="col-md-2">
                 <strong>Short URL:</strong>
             </span>
-            <a asp-controller="ShortUrls" asp-action="RedirectTo" asp-route-path="@ViewData["Path"]">
-                http://localhost:5000/ShortUrls/RedirectTo/@ViewData["Path"]
+            <a asp-controller="ShortUrlRedirect" asp-action="RedirectTo" asp-route-path="@ViewData["Path"]">
+                http://localhost:5000/@ViewData["Path"]
             </a>
         </li>
     </ul>
